chess: widen hot-path score variables to native int for eZ80
On eZ80, int (24-bit) is native while int16_t requires extra cycles.
Widen all score-related local variables, function parameters, and
return types in search/eval/TT to native int. Eliminates int32_t
from tapered eval (mg*phase fits in 24-bit).

Benchmark: eval -5.5% cy/call, overall -2.6% cy/node, binary -514B.
History table and scored_move_t.score kept as int16_t (array access
penalty outweighs ALU benefit for wide data structures).

+++ chess/engine/src/eval.c
@@ -274,11 +274,11 @@ static void build_pawn_info(const board_t *b,
 
 /* ========== Main Evaluation ========== */
 
-int16_t evaluate(const board_t *b)
+int evaluate(const board_t *b)
 {
     static uint8_t pawn_atk[128];
-    int16_t mg, eg, phase;
-    int32_t score;
+    int mg, eg, phase;
+    int score;
     uint8_t i, sq, row, col, type;
     uint8_t w_pawns[8], b_pawns[8];
     EP_VARS;
@@ -584,8 +584,8 @@ int16_t evaluate(const board_t *b)
     if (phase > PHASE_MAX) phase = PHASE_MAX;
     if (phase < 0) phase = 0;
 
-    score = ((int32_t)mg * phase + (int32_t)eg * (PHASE_MAX - phase)) / PHASE_MAX;
+    score = (mg * phase + eg * (PHASE_MAX - phase)) / PHASE_MAX;
 
     /* Return from side-to-move perspective */
-    return (b->side == WHITE) ? (int16_t)score : (int16_t)(-score);
+    return (b->side == WHITE) ? score : -score;
 }
